owrt: targets: Remove pcie_bus_config performance
Change is to configure QC RC MPS 128 bytes as 8168 device only supports
128 bytes. Without this, since RC is using pcie_bus_perf flag
MOS is configured to 256 bytes, when EP MRRS is changed to 4096 bytes
RC is responding with packets of 256 bytes causing functionality break.

Since pcie_bus_perf is used for many products, this change provides the
provision to configure RC to 128 bytes for realtek 8168 attach alone.

Change-Id: I10cf318bda23b37814e6b65c9da0bc932033754a